[Chore] #146 Goal 내용을 ParentGoal 파일로 이동
